requirements.txt
fastapi
uvicorn
redis
bcrypt
requests
python-dotenv